added postgres database - moved scraper service to api directory - added airflow to main docker compose - removed unnecessary files and directories - renamed scraper service class
