+++ api/src/di/dependencies.py
@@ -7,7 +7,7 @@ from riot_games.service import RiotGamesService
 import os
 from sqlalchemy import create_engine, MetaData
 from reporter.service import LoLStatsService
-from scraper.service import LeagueOfGraphsScraperService
+from scraper.service import ScraperService
 
 #Buckets
 account_by_id_limiter = AsyncLeakyBucket(capacity=1000,leak_rate=1000/60) # 2000 requests every 1 minute
@@ -17,6 +17,9 @@ match_timeline_by_match_id_limiter = AsyncLeakyBucket(2000,2000/10) # 2000 reque
 matches_by_puuid_limiter = AsyncLeakyBucket(2000,2000/10) # 2000 requests every 10 seconds
 
 riot_games_service = RiotGamesService(os.getenv("RIOT_API_KEY"))
+
+_postgres_engine = None
+_postgres_metadata = None
 
 #Dependencies for LolStatsService
 def provide_mongo_client():
@@ -37,7 +40,6 @@ def provide_lol_stats_service(
 def provide_riot_games_service():
     yield riot_games_service
 
-
 #Dependency for RiotGames
 async def acquire_account_puuid_limiter(request: Request):
     await account_by_puuid_limiter.acquire()
@@ -54,15 +56,8 @@ async def acquire_match_timeline_by_match_id_limiter(request: Request):
 async def acquire_matches_by_puuid_limiter(request: Request):
     await matches_by_puuid_limiter.acquire()
 
-def provide_scraper():
-    scraper = LeagueOfGraphsScraperService()
-    
-scraper_service = None
-_postgres_engine = None
-_postgres_metadata = None
-
 def provide_scraper_service():
-    scraper_service = LeagueOfGraphsScraperService()
+    scraper_service = ScraperService()
 
     yield scraper_service
     # anything that needs to be executed after the function that's being injected with this dependency should go here
@@ -81,11 +76,6 @@ def provide_postgres_engine():
     
     yield _postgres_engine
 
-def provide_mongo_client():
-    mongo_client = MongoClient(host=os.environ['MONGODB_URL'], port=int(os.environ['MONGODB_PORT']))
-    
-    yield mongo_client
-    
 # Dependency function for LoLStatsService
 def provide_lol_stats_service(
     mongo_client: Annotated[MongoClient, Depends(provide_mongo_client)]
