+++ api/src/riot_games/service.py
@@ -1,4 +1,3 @@
-import os
 import requests
 
 
@@ -22,7 +21,7 @@ class RiotGamesService():
 
     def get_account_by_puuid(self, puuid: str, region: str='europe'):
         base_url = self.base_url.replace(self.region_placeholder, region)
-        url = f'https://{base_url}/riot/account/v1/accounts/by-puuid/{puuid}'
+        url = f'{base_url}/riot/account/v1/accounts/by-puuid/{puuid}'
         
         response = requests.get(url,headers=self.headers)
         if "status" in response.json().keys():
@@ -31,7 +30,7 @@ class RiotGamesService():
 
     def get_match_by_match_id(self, match_id: str, region: str='europe'):
         base_url = self.base_url.replace(self.region_placeholder, region)
-        url = f'https://{base_url}/lol/match/v5/matches/{match_id}'
+        url = f'{base_url}/lol/match/v5/matches/{match_id}'
         
         response = requests.get(url,headers=self.headers)
         if "status" in response.json().keys():
@@ -41,7 +40,7 @@ class RiotGamesService():
     
     def get_matches_by_puuid(self,puuid: str, region: str='europe'):
         base_url = self.base_url.replace(self.region_placeholder,region)
-        url = f'https://{base_url}/lol/match/v5/matches/by-puuid/{puuid}/ids'
+        url = f'{base_url}/lol/match/v5/matches/by-puuid/{puuid}/ids'
         
         response = requests.get(url, headers=self.headers)
         if "status" in response.json().keys():
@@ -50,7 +49,7 @@ class RiotGamesService():
     
     def get_match_timeline_by_match_id(self, match_id: str, region: str='europe'):
         base_url = self.base_url.replace(self.region_placeholder,region)
-        url = f'https://{base_url}/lol/match/v5/matches/{match_id}/timeline'
+        url = f'{base_url}/lol/match/v5/matches/{match_id}/timeline'
         
         response = requests.get(url, headers=self.headers)
         if "status" in response.json().keys():
