+++ api/src/scraper/exceptions.py
