+++ api/src/scraper/models.py
