+++ api/src/scraper/router.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter, Query, Depends
 from typing import Annotated
 
-from .service import LeagueOfGraphsScraper
-from .models import LeagueOfGraphsModel
+from scraper.service import ScraperService
+from scraper.models import LeagueOfGraphsModel
 
-from di.dependencies import provide_scraper
+from di.dependencies import provide_scraper_service
 
 
 router = APIRouter(
@@ -12,11 +12,11 @@ router = APIRouter(
 )
 
 # defining an annotated type to reuse in multiple endpoints
-Scraper = Annotated[LeagueOfGraphsScraper, Depends(provide_scraper)]
+Scraper = Annotated[ScraperService, Depends(provide_scraper_service)]
 
 @router.get('')
 async def scrape_data(model: Annotated[LeagueOfGraphsModel, Query()], scraper: Scraper):
-    return await scraper.get_data_from_page(
+    return await scraper.scrape_pages(
         start_page=model.start_page,
         end_page=model.end_page,
         region=model.region
