--- infra/scraper/requirements.txt
@@ -1,8 +1,0 @@
-beautifulsoup4==4.13.3
-fastapi[standard]==0.115.13
-requests
-lxml[html_clean]
-requests-html
-pyppeteer
-cloudscraper
-aiohttp
